дз

+++ src/main/java/HomeworkSkypro/calculator/CalculatorService.java
@@ -2,8 +2,8 @@ package HomeworkSkypro.calculator;
 
 public interface CalculatorService {
     String hello();
-    double plus(double num1, double num2);
-    double minus(double num1,double num2);
-    double multiply(double num1,double num2);
-    double divide(double num1,double num2);
+    Integer plus(Integer num1, Integer num2);
+    Integer minus(Integer num1,Integer num2);
+    Integer multiply(Integer num1,Integer num2);
+    int divide(int num1,int num2);
 }
